+++ include/H2DE/H2DE_object.h
@@ -15,6 +15,7 @@ private:
     H2DE_Engine* engine;
     std::vector<H2DE_LevelObject*>* objects;
     H2DE_LevelObjectData data;
+    bool hidden = false;
 
     void snap(H2DE_LevelRect thisRect, H2DE_LevelRect otherRect, H2DE_Face face);
 
@@ -30,7 +31,27 @@ public:
      * \returns level object's data
      * \since H2DE-2.0.5
      */
-    friend H2DE_LevelObjectData* H2DE_GetObjectData(H2DE_LevelObject* object);
+    friend H2DE_LevelObjectData* H2DE_GetLevelObjectData(H2DE_LevelObject* object);
+    /**
+     * Shows a level object
+     * \param object pointer to a level object
+     * \since H2DE-2.1.7
+     */
+    friend void H2DE_ShowLevelObject(H2DE_LevelObject* object);
+    /**
+     * Hides a level object
+     * \param object pointer to a level object
+     * \since H2DE-2.1.7
+     */
+    friend void H2DE_HideLevelObject(H2DE_LevelObject* object);
+
+    /**
+     * Checks if a level object is hidden
+     * \param object pointer to a level object
+     * \return true if the level object is hidden
+     * \since H2DE-2.1.7
+     */
+    friend bool H2DE_IsLevelObjectHidden(H2DE_LevelObject* object);
 };
 
 #endif